mod-ur: separate spawn control from @api-cli.mts as serve-control
